add navbar action to prevent refresh lost
