+++ front_end/src/components/layout/Navbar.js
@@ -2,7 +2,7 @@ import React, { Component } from "react";
 import { Link } from "react-router-dom";
 import PropTypes from "prop-types";
 import { connect } from "react-redux";
-import { logoutUser,setCurrentUser } from "../../actions/authActions";
+import { logoutUser,getCurrentUser } from "../../actions/authActions";
 import { clearCurrentProfile } from "../../actions/profileActions";
 import AnchorLink from "react-anchor-link-smooth-scroll";
 import NavbarMenu from "./NavbarMenu";
@@ -29,26 +29,13 @@ class Navbar extends Component {
   componentWillMount(){
     if (localStorage.jwtToken) {
       const decode = jwt_decode(localStorage.jwtToken);
-      axios.get("api/users/current").then(res =>{
-        console.log(res.data)
-        const decoded = {
-          id: decode.id,
-          iat: decode.iat,
-          exp:decode.exp,
-          firstname:res.data.firstName,
-          lastname:res.data.lastName,
-          email:res.data.email,
-          rewardPoints:res.data.rewardPoints
-        }
-        this.props.setCurrentUser(decoded)
-      })
+      this.props.getCurrentUser(decode);
     }
   }
 
 
   render() {
     // Markup shown on the right hand side of Navbar when user is GUEST.
-    console.log(this.props)
     let guestMarkUp = (
       <Grid
         container
@@ -182,5 +169,5 @@ let mapStateToProps = state => ({
 
 export default connect(
   mapStateToProps,
-  { logoutUser, clearCurrentProfile,setCurrentUser }
+  { logoutUser, clearCurrentProfile,getCurrentUser }
 )(Navbar);
